Correctly Send Normal Data to GPU - Fixes the way GLRenderer sends normal data to the GPU. - Fixes COLLADA loading so that it correctly assembles vertex data.

+++ lib/polygon_rs/src/geometry/mesh.rs
@@ -31,7 +31,7 @@ impl Mesh {
             },
             normal_attribute: VertexAttribute {
                 stride: 3,
-                offset: positions_raw.len() as u32,
+                offset: positions_raw.len(),
             },
         }
     }
@@ -39,6 +39,6 @@ impl Mesh {
 
 #[derive(Debug, Clone, Copy)]
 pub struct VertexAttribute {
-    pub stride: u32,
-    pub offset: u32,
+    pub stride: usize,
+    pub offset: usize,
 }
